Expense app with user Authentication

+++ 4_Expense Tracker - Node.js Project/Backend/controllers/expenseControllers.js
@@ -20,7 +20,8 @@ exports.addExpense = async (req, res) => {
         // 400 Bad Request response status code indicates that the server cannot or will not process the request due to something that is perceived to be a client error (for example, malformed request syntax, invalid request message framing, or deceptive request routing).
     }
 
-    await Expense.create({ expenseamount, description, category }).then((expense) => {
+    // req.user.addExpense({ expenseamount, description, category }).then((expense) => {    // // Magic Function
+    await Expense.create({ expenseamount, description, category, userId: req.user.id }).then((expense) => {
         return res.status(201).json({ expense, success: true, message: "Expense Added to DB" });
         // 201 Created success status response code indicates that the request has succeeded and has led to the creation of a resource.
       }
@@ -34,7 +35,8 @@ exports.addExpense = async (req, res) => {
 // Get Expense
 exports.getExpense = (req, res) => {
   try{
-  Expense.findAll().then(expenses => {
+  // req.user.getExpense().then(expenses => {   // // Magic Function 
+  Expense.findAll({ where : {userId: req.user.id}}).then(expenses => {
     return res.status(200).json({expenses, success:true})
   })
   }catch(err) {
@@ -52,7 +54,10 @@ exports.deleteExpense = (req, res) => {
     // 400 Bad Request response status code indicates that the server cannot or will not process the request due to something that is perceived to be a client error (for example, malformed request syntax, invalid request message framing, or deceptive request routing).
   
   }
-  Expense.destroy({ where: { id: expenseid }}).then(() => {
+  Expense.destroy({ where: { id: expenseid, userId: req.user.id }}).then((noOfRows) => {
+    if(noOfRows === 0){
+      return res.status(404).json({success: false, message: 'Expense does not belog to the user'})
+    }
     return res.status(200).json({ success: true, message: 'Deleted Successfully'})
   }).catch(err => {
     console.log(err);
